# Projects/nexus-os/packages/schemas/openapi.json
{
  "components": {
    "schemas": {
      "AgentCreate": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "fallback_models": {
            "items": {
              "type": "string"
            },
            "maxItems": 6,
            "title": "Fallback Models",
            "type": "array"
          },
          "icon": {
            "default": "bot",
            "maxLength": 40,
            "title": "Icon",
            "type": "string"
          },
          "knowledge_sources": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Knowledge Sources",
            "type": "array"
          },
          "max_runtime_s": {
            "default": 300,
            "maximum": 3600.0,
            "minimum": 10.0,
            "title": "Max Runtime S",
            "type": "integer"
          },
          "max_steps": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Steps",
            "type": "integer"
          },
          "max_tool_calls": {
            "default": 40,
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Max Tool Calls",
            "type": "integer"
          },
          "memory_scope": {
            "$ref": "#/components/schemas/MemoryScope"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/AgentPermissions"
          },
          "preferred_model": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Model"
          },
          "reasoning_mode": {
            "default": "off",
            "enum": [
              "off",
              "light",
              "deep"
            ],
            "title": "Reasoning Mode",
            "type": "string"
          },
          "role": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "system_prompt": {
            "default": "",
            "maxLength": 20000,
            "title": "System Prompt",
            "type": "string"
          },
          "temperature": {
            "default": 0.2,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "token_budget": {
            "default": 200000,
            "maximum": 5000000.0,
            "minimum": 1000.0,
            "title": "Token Budget",
            "type": "integer"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "maxItems": 60,
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "name",
          "role"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
      "AgentDefinition": {
        "properties": {
          "builtin": {
            "default": false,
            "title": "Builtin",
            "type": "boolean"
          },
          "color": {
            "default": "",
            "title": "Color",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "fallback_models": {
            "items": {
              "type": "string"
            },
            "title": "Fallback Models",
            "type": "array"
          },
          "icon": {
            "default": "bot",
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "knowledge_sources": {
            "items": {
              "type": "string"
            },
            "title": "Knowledge Sources",
            "type": "array"
          },
          "max_runtime_s": {
            "default": 300,
            "title": "Max Runtime S",
            "type": "integer"
          },
          "max_steps": {
            "default": 20,
            "title": "Max Steps",
            "type": "integer"
          },
          "max_tool_calls": {
            "default": 40,
            "title": "Max Tool Calls",
            "type": "integer"
          },
          "memory_scope": {
            "$ref": "#/components/schemas/MemoryScope"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "$ref": "#/components/schemas/AgentPermissions"
          },
          "preferred_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Model"
          },
          "reasoning_mode": {
            "default": "off",
            "enum": [
              "off",
              "light",
              "deep"
            ],
            "title": "Reasoning Mode",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "default": "idle",
            "enum": [
              "idle",
              "busy",
              "disabled"
            ],
            "title": "Status",
            "type": "string"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          },
          "temperature": {
            "default": 0.2,
            "title": "Temperature",
            "type": "number"
          },
          "token_budget": {
            "default": 200000,
            "title": "Token Budget",
            "type": "integer"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "title": "Tools",
            "type": "array"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "role"
        ],
        "title": "AgentDefinition",
        "type": "object"
      },
...
      "AgentRunOut": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "objective_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective Id"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "step_count": {
            "title": "Step Count",
            "type": "integer"
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "tokens_in": {
            "title": "Tokens In",
            "type": "integer"
          },
          "tokens_out": {
            "title": "Tokens Out",
            "type": "integer"
          },
          "tool_call_count": {
            "title": "Tool Call Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "agent_id",
          "project_id",
          "objective_id",
          "task_id",
          "status",
          "model",
          "step_count",
          "tool_call_count",
          "tokens_in",
          "tokens_out",
          "attempt",
          "result",
          "error",
          "started_at",
          "finished_at"
        ],
        "title": "AgentRunOut",
        "type": "object"
      },
      "AgentRunRequest": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "description": "Agent id or slug",
            "maxLength": 60,
            "minLength": 1,
            "title": "Agent",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Manual model override: '<provider_id>:<model>'",
            "title": "Model"
          },
          "private": {
            "default": false,
            "title": "Private",
            "type": "boolean"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "prompt": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "agent",
          "project_id",
          "prompt"
        ],
        "title": "AgentRunRequest",
        "type": "object"
      },
...
      "AnswerBody": {
        "additionalProperties": false,
        "properties": {
          "text": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "AnswerBody",
        "type": "object"
      },
      "ApprovalDecision": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "approve_once",
              "approve_session",
              "deny"
            ],
            "title": "Decision",
            "type": "string"
          },
          "edited_arguments": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Arguments"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "decision"
        ],
        "title": "ApprovalDecision",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "NEXUS OS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents": {
      "get": {
        "operationId": "list_agents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentDefinition"
                  },
                  "title": "Response List Agents",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run": {
      "post": {
        "operationId": "run_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/{agent_id}": {
      "delete": {
        "operationId": "delete_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "update_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDefinition"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/approvals": {
      "get": {
        "operationId": "list_approvals",
        "parameters": [
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalOut"
                  },
                  "title": "Response List Approvals",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Approvals",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/approvals/grants/{grant_id}": {
      "delete": {
        "operationId": "revoke_grant",
        "parameters": [
          {
            "in": "path",
            "name": "grant_id",
            "required": true,
            "schema": {
              "title": "Grant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Grant",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{approval_id}": {
      "get": {
        "operationId": "get_approval",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Approval",
        "tags": [
          "approvals"
        ]
      }
    },
    "/api/approvals/{approval_id}/decision": {
      "post": {
        "operationId": "decide",
        "parameters": [
          {
            "in": "path",
            "name": "approval_id",
            "required": true,
            "schema": {
              "title": "Approval Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "approvals"
        ]
      }
    },
...
    "/api/runs": {
      "get": {
        "operationId": "list_runs",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          },
          {
            "in": "query",
            "name": "agent",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent"
            }
          },
          {
            "in": "query",
            "name": "status_filter",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRunOut"
                  },
                  "title": "Response List Runs",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "get_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/runs/{run_id}/answer": {
      "post": {
        "operationId": "answer_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Answer Run",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/runs/{run_id}/resume": {
      "post": {
        "operationId": "resume_run",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Run",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}